test: +771 tests, core type fixes, coverage boost (#130)
* test: add 394 tests across core, execution, backtest — coverage boost

Comprehensive test coverage for previously untested modules:
- core/indicators: kuramoto, entropy (phase sync, Shannon entropy)
- core/utils: metrics (Prometheus collector, fallback quantiles)
- core/neuro: integrated, training, behavioral_profiler
- core/physics, core/ml, core/security, core/compliance
- core/architecture: system_principles
- execution: live_loop, order_lifecycle, oms, risk/core
- backtest: engine, event_driven

All 394 tests pass. Covers ~15,000 lines of previously untested source code.


* test: add analytics/irreversibility, data/feature_store, events/sourcing tests

29 additional tests covering:
- IGSConfig and IGSStreamer (irreversibility signals)
- RetentionPolicy and feature store exceptions
- EventEnvelope, AggregateSnapshot, AggregateRoot


* style: fix ruff lint and black formatting across all new test files

- Remove unused imports (F401)
- Fix lambda assignments (E731) → def functions
- Fix semicolons (E702) → separate statements
- Apply black formatting
- All 423 tests pass, all lint checks clean


* test: add LiveLoopConfig, RangeCheck, SignalFilter tests — coverage push to 89%

59 additional tests:
- LiveLoopConfig: 18 tests covering all config validation
- RangeCheck/TemporalContract: 15 tests for quality gate dataclasses
- SignalFilter: 26 tests for filter_invalid_values, FilterResult, config


* test: add RiskLimits, KillSwitchStateRecord, RiskStateRecord, LiveLoopConfig, SignalFilter tests

103 additional tests covering key execution risk infrastructure:
- RiskLimits: 18 parametrized tests (drawdown conversion, clamping)
- KillSwitchStateRecord: 11 tests (validation, coercion, timestamps)
- RiskStateRecord: 9 tests (sanitization, type checks)
- LiveLoopConfig: 18 tests (all config validation paths)
- RangeCheck/TemporalContract: 15 tests
- SignalFilter: 26 tests (invalid values, strategies, DoS protection)
- FilterResult: 6 tests


* test: add 31 transaction cost model tests for backtest engine

Covers TransactionCostModel, PerUnitCommission, FixedBpsCommission,
PercentVolumeCommission, FixedSpread, BpsSpread, FixedSlippage,
VolumeProportionalSlippage, SquareRootSlippage, BorrowFinancing.


* test: add 131 tests for pipeline, backfill, portfolio, circuit breaker, ledger, watchdog, sizer

Major coverage push across execution and core/data subsystems:
- execution.position_sizer: 12 tests (risk sizing, leverage caps, validation)
- execution.portfolio: 14 tests (PnL, snapshots, fills, fees)
- execution.resilience.circuit_breaker: 17 tests (states, recovery, failure rate)
- execution.order_ledger: 25 tests (_coerce, config, LedgerMetadata)
- execution.watchdog: 14 tests (Watchdog, _WorkerSpec, config clamping)
- core.data.pipeline: 24 tests (all dataclasses, SLA, anonymization, exceptions)
- core.data.backfill: 16 tests (CacheKey, CacheEntry, LayerCache)

Total: 131 tests, all passing, lint clean.


* test: add 29 tests for core.neuro.ecs_regulator — validation paths

Covers ECSInspiredRegulator initialization validation for all parameters,
StressMode enum, ECSMetrics, StabilityMetrics dataclasses.


* test: add 21 tests for async_ingestion _TickMetricBatcher

Covers threshold clamping, flush logic, context manager,
exception handling, parametrized threshold configs.


* test: add 33 tests for advanced_risk_manager and DataPipeline init

- advanced_risk_manager: 24 tests (StressResponseProtocol, RiskState ordering,
  MarketDepthData mid/spread, LiquidityMetrics, FreeEnergyState)
- DataPipeline init: 9 tests (config validation, resolve_schema/gate, result)


* style: apply black formatter to all new test files (CI compliance)

17 files reformatted to satisfy python-quality gate.
All 771 tests still pass. Ruff check clean.


* fix(types): zero mypy errors across PR + clean up core type annotations

## Tests
Replaced `X = None` fallback pattern in try/except ImportError blocks with
`pytest.skip(..., allow_module_level=True)`. mypy cannot reconcile
`type[X] = None` assignments; module-level skip is the canonical idiom.

Files fixed:
- tests/core/test_compliance_modules.py
- tests/core/test_ml_modules.py
- tests/core/test_physics_modules.py
- tests/core/test_security_modules.py
- tests/execution/test_risk_core.py
- tests/analytics/test_irreversibility.py (removed broken IGSStreamer block)

## Core (pre-existing mypy errors surfaced by new test imports)

**core/kuramoto/delayed.py**
- `history_fn: callable | None` → `Callable[[float], NDArray[np.float64]] | None`
- Cast `tau` scalar path to float with `# type: ignore[arg-type]` for numpy scalar.

**core/kuramoto/ricci_flow_engine.py**
- Optional Ollivier/Forman import fallback annotated with `Any` for `_lite_ricci`.
- `_to_light_graph` local `lite` variable typed as `Any` to reconcile the
  LightGraph / _SimpleUndirectedGraph fallback branch.

**core/physics/higher_order_kuramoto.py**
- `omega` / `theta0` Optional parameters resolved into typed local arrays
  (`omega_arr`, `theta0_arr`) via `np.asarray(..., dtype=np.float64)` so the
  downstream RK4 call receives a concrete `NDArray[np.float64]`.
- `corr` / return path likewise normalised through `np.asarray`.

**core/physics/uncertainty.py**
- Cast `heisenberg_uncertainty` result to `float` (it may return numpy scalar).
- `violation_factor` and `is_valid` explicitly cast to Python built-ins.

**core/physics/forman_ricci.py** / **liquidity_coupling.py** / **wave_propagation.py**
- Narrow numpy return types via `np.asarray(..., dtype=np.float64)` and
  `.astype(np.float64)` where the upstream call produced a union type.

## Result

- 0 mypy errors across 37 PR source files.
- 0 ruff warnings.
- `black --check` passes.
- 771 tests pass, 1 skipped (behavioural profiler optional dep).


* style: black --target-version py311 for modified core files

CI runs Python 3.11; local env is 3.12 which wraps lines slightly
differently. Re-format the 7 files modified in the previous commit with
--target-version py311 so CI's black --check matches byte-for-byte.


---------

+++ core/physics/wave_propagation.py
@@ -164,9 +164,7 @@ class GraphDiffusionEngine:
             return [asset_names[i] for i in indices]
         return indices.tolist()
 
-    def laplacian_eigenvalues(
-        self, L: NDArray[np.float64]
-    ) -> NDArray[np.float64]:
+    def laplacian_eigenvalues(self, L: NDArray[np.float64]) -> NDArray[np.float64]:
         """Compute eigenvalues of graph Laplacian.
 
         For valid Laplacians:
@@ -175,7 +173,7 @@ class GraphDiffusionEngine:
         - Second smallest = algebraic connectivity (Fiedler value)
         """
         eigenvalues = np.linalg.eigvalsh(L)
-        return eigenvalues
+        return eigenvalues.astype(np.float64)
 
 
 __all__ = ["GraphDiffusionEngine"]
